+++ src/app/app.ts
@@ -57,7 +57,7 @@ creator
 // Log client errors
 client.on('error', Log.error);
 
-client.login(process.env.DISCORD_BOT_TOKEN).catch(Log.debug);
+client.login(process.env.DISCORD_BOT_TOKEN).catch(Log.error);
 
 function initializeClient(): Client {
 	const clientOptions: ClientOptions = {